Generalize container of statistics algorithms.
As written, the statistics algorithms only work on standard lists.
However, Haskell lists are very inefficient and most production code
uses `Vector` or `Array`. We can generalize these functions by mostly
changing the type signature.

Added a few more methods for median due to the fact that this algorithm
cannot be generalized due to different sorting implementations. However,
this also results in adding functions for finding the median of already
sorted arrays which can result in better complexity if the input is
guaranteed to be sorted.